pyproject.toml
[tool.poetry]
name = "gandalf_bot"
version = "0.1.0"
description = ""
authors = ["Celeo <mattboulanger@fastmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.5.1"
loguru = "^0.5.3"
httpx = "^0.16.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
